extract smart agent logic to a new commander class
